+++ src/cliness2d_test.cc
@@ -39,7 +39,7 @@ TEST_CASE("LinesS2D: horizontal segments along y=cy are recovered") {
   REQUIRE(finder->nresults() >= 1);
   const float theta = finder->angle(0);
   const float offset = finder->offset(0);
-  CHECK(std::cos(theta - M_PI_2) == doctest::Approx(1.0).epsilon(0.05));
+  CHECK(std::cos(theta - static_cast<float>(M_PI_2)) == doctest::Approx(1.0).epsilon(0.05));
   CHECK(offset == doctest::Approx(cy).epsilon(0.5));
 }
